Add files via upload
New ones + minor fixes

-Improved some white spines and sega logos for a better, more cohesive and smooth scrolling 
-Fixed D2 upscaling artifacs
-Minor fixes in Marvel VS Capcom 2 , Better colors in Airforce Delta, and some more changes i can't remember

With this commit we almost have a complete USA boxart collection.
The only 2 boxarts still missing are:

-Grand Theft Auto 2 
-POD - Speedzone

i've searched all over the internet for a good guality scans of these two, to work with, but no luck.  (Maybe one kind soul, who has a physical copy of these can provide a better scan. Mabye i will open a reddit for that. And maybe the GTA 2 cover could be made from scratch. In the end it´s only some logos and a plain backround).